Instantly update the feed after a user posts
