+++ prototypes/web/postboard/frontend/src/app/types.tsx
@@ -1,0 +1,5 @@
+export type FeedData = Array<{
+  author: string;
+  content: string;
+  created: string;
+}>;
